Completed the first step in the play
After Deal allows the top card of player hand to be revealed and the player to choose which property they wish to "bet" on. The computer works out what the bet is worth and then checks the value of the computer hand top card and evaluates it. Next step what to do with cards - if the values are the same. If they are different it is easy!
If player wins another go will happen otherwise play turns over to the computer!